+++ data/service/api/v1/data_set.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/tidepool-org/platform/data"
 	dataService "github.com/tidepool-org/platform/data/service"
-	"github.com/tidepool-org/platform/errors"
 	"github.com/tidepool-org/platform/page"
 	"github.com/tidepool-org/platform/request"
 	"github.com/tidepool-org/platform/user"
@@ -52,7 +51,7 @@ func ListUserDataSets(dataServiceContext dataService.Context) {
 	if !details.IsService() && details.UserID() != userID {
 		permissions, err := dataServiceContext.UserClient().GetUserPermissions(req.Context(), details.UserID(), userID)
 		if err != nil {
-			if errors.Code(err) == request.ErrorCodeUnauthorized {
+			if request.IsErrorUnauthorized(err) {
 				responder.Error(http.StatusForbidden, request.ErrorUnauthorized())
 			} else {
 				responder.Error(http.StatusInternalServerError, err)