+++ data/service/api/v1/datasets_delete.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 
 	dataService "github.com/tidepool-org/platform/data/service"
-	"github.com/tidepool-org/platform/errors"
 	"github.com/tidepool-org/platform/log"
 	"github.com/tidepool-org/platform/request"
 	"github.com/tidepool-org/platform/service"
@@ -43,7 +42,7 @@ func DatasetsDelete(dataServiceContext dataService.Context) {
 		var permissions user.Permissions
 		permissions, err = dataServiceContext.UserClient().GetUserPermissions(ctx, authUserID, *targetUserID)
 		if err != nil {
-			if errors.Code(err) == request.ErrorCodeUnauthorized {
+			if request.IsErrorUnauthorized(err) {
 				dataServiceContext.RespondWithError(service.ErrorUnauthorized())
 			} else {
 				dataServiceContext.RespondWithInternalServerFailure("Unable to get user permissions", err)