+++ request/errors_test.go
@@ -30,6 +30,19 @@ var _ = Describe("Errors", func() {
 		})
 	})
 
+	Context("ErrorInternalServerError", func() {
+		It("returns the expected error", func() {
+			cause := errors.New("error")
+			err := request.ErrorInternalServerError(cause)
+			Expect(err).ToNot(BeNil())
+			Expect(errors.Code(err)).To(Equal("internal-server-error"))
+			Expect(errors.Cause(err)).To(Equal(cause))
+			bytes, bytesErr := json.Marshal(errors.Sanitize(err))
+			Expect(bytesErr).ToNot(HaveOccurred())
+			Expect(bytes).To(MatchJSON(`{"code": "internal-server-error", "title": "internal server error", "detail": "internal server error"}`))
+		})
+	})
+
 	DescribeTable("have expected details when error",
 		errorsTest.ExpectErrorDetails,
 		Entry("is ErrorTooManyRequests", request.ErrorTooManyRequests(), "too-many-requests", "too many requests", "too many requests"),
@@ -57,5 +70,61 @@ var _ = Describe("Errors", func() {
 			Entry("is another error", errors.New("test-error"), 500),
 			Entry("is nil error", nil, 500),
 		)
+	})
+
+	Context("IsErrorInternalServerError", func() {
+		It("returns false if the error does not have a code", func() {
+			Expect(request.IsErrorInternalServerError(errors.New("error"))).To(BeFalse())
+		})
+
+		It("returns false if the error code is not ErrorCodeInternalServerError", func() {
+			Expect(request.IsErrorInternalServerError(request.ErrorUnauthenticated())).To(BeFalse())
+		})
+
+		It("returns true if the error code is ErrorCodeInternalServerError", func() {
+			Expect(request.IsErrorInternalServerError(request.ErrorInternalServerError(errors.New("error")))).To(BeTrue())
+		})
+	})
+
+	Context("IsErrorUnauthenticated", func() {
+		It("returns false if the error does not have a code", func() {
+			Expect(request.IsErrorUnauthenticated(errors.New("error"))).To(BeFalse())
+		})
+
+		It("returns false if the error code is not ErrorCodeUnauthenticated", func() {
+			Expect(request.IsErrorUnauthenticated(request.ErrorInternalServerError(errors.New("error")))).To(BeFalse())
+		})
+
+		It("returns true if the error code is ErrorCodeUnauthenticated", func() {
+			Expect(request.IsErrorUnauthenticated(request.ErrorUnauthenticated())).To(BeTrue())
+		})
+	})
+
+	Context("IsErrorUnauthorized", func() {
+		It("returns false if the error does not have a code", func() {
+			Expect(request.IsErrorUnauthorized(errors.New("error"))).To(BeFalse())
+		})
+
+		It("returns false if the error code is not ErrorCodeUnauthorized", func() {
+			Expect(request.IsErrorUnauthorized(request.ErrorInternalServerError(errors.New("error")))).To(BeFalse())
+		})
+
+		It("returns true if the error code is ErrorCodeUnauthorized", func() {
+			Expect(request.IsErrorUnauthorized(request.ErrorUnauthorized())).To(BeTrue())
+		})
+	})
+
+	Context("IsErrorResourceNotFound", func() {
+		It("returns false if the error does not have a code", func() {
+			Expect(request.IsErrorResourceNotFound(errors.New("error"))).To(BeFalse())
+		})
+
+		It("returns false if the error code is not ErrorCodeResourceNotFound", func() {
+			Expect(request.IsErrorResourceNotFound(request.ErrorInternalServerError(errors.New("error")))).To(BeFalse())
+		})
+
+		It("returns true if the error code is ErrorCodeResourceNotFound", func() {
+			Expect(request.IsErrorResourceNotFound(request.ErrorResourceNotFound())).To(BeTrue())
+		})
 	})
 })