+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD
 
+* Add shortcuts for request error comparisons
 * Add streamed response from client in addition to existing data response
 * Add ability to specify authorization mechanism at client creation
 * Improve client tests